fix: fetch full name from (authenticated) root layout
