+++ app/(authenticated)/_layout.tsx
@@ -11,10 +11,12 @@ import Colors from "@/constants/Colors";
 import { BlurView } from "expo-blur";
 import { supabase } from "@/utils/supabase";
 import { useUserIdStore } from "@/utils/store/user-id-store";
+import { useProfileStore } from "@/utils/store/profile-store";
 
 const Layout = () => {
   const segments = useSegments();
   const setUserId = useUserIdStore((state) => state.setUserId);
+  const setFullName = useProfileStore((state) => state.setFullName);
 
   useEffect(() => {
     const getUserId = async () => {
@@ -27,7 +29,22 @@ const Layout = () => {
       }
       setUserId(user.id);
     };
-    getUserId();
+
+    const getFullName = async () => {
+      const { data, error } = await supabase
+        .from("profiles")
+        .select("full_name")
+        .single();
+      if (error) {
+        setFullName("User");
+        return;
+      }
+      setFullName(data.full_name);
+    };
+
+    getUserId().then(() => {
+      getFullName();
+    });
   }, []);
 
   return (
